dlna: set more correct mime type for other resources
The code currently hardcodes `text/srt` for all subtitles.

`text/srt` is wrong, it seems `application/x-subrip` is the official
extension coming from the official mime database. Use that one in the
tests, and also fix `fs/mimetype.go`.

Fix the mime type for non-.srt subtitles too, by updating this code to
pick the mime type similar to how the mime type of the main media file
is determined, and include a testcase with .idx/.sub.